cephadm: move packager types and funcs to packagers.py
Pair-programmed-with: Adam King <adking@redhat.com>
